Move playlistTrack to new package
