+++ src/main/java/tech/makers/aceplay/playlist_track/PlaylistTrack.java
@@ -1,5 +1,6 @@
-package tech.makers.aceplay.playlist;
+package tech.makers.aceplay.playlist_track;
 
+import tech.makers.aceplay.playlist.Playlist;
 import tech.makers.aceplay.track.Track;
 
 import javax.persistence.*;
