+++ src/main/java/tech/makers/aceplay/playlist_track/PlaylistTrackRepository.java
@@ -1,6 +1,7 @@
-package tech.makers.aceplay.playlist;
+package tech.makers.aceplay.playlist_track;
 
 import org.springframework.data.repository.CrudRepository;
+import tech.makers.aceplay.playlist.Playlist;
 import tech.makers.aceplay.track.Track;
 
 import java.util.Optional;
